Fix 0px height issue on first layout
When the ViewPager is first layed out onMeasure can be called before any children have been added (I suspect this when the viewpager is in a scrollview). This results in the viewpager having no height. onMeasure needs to be called at the start so that it can add the child views (private method populate() is called in the parent onMeasure)